Move `Address`, `Hash`, `TxIn`, `TxOut` to test suite.
These types are no longer used in the public API, but they are still
used within the test suite, for convenience.
